editing backend

+++ backend/src/controllers/list.controller.js
@@ -2,10 +2,11 @@ import boardSchema from "../models/boardSchema.js";
 import listSchema from "../models/listSchema.js";
 import taskSchema from "../models/taskSchema.js";
 import commentSchema from "../models/commentSchema.js";
+import { order_resolver } from "../utils/order_resolver.js";
 const AddList= async(req,res,next)=>{
     const {board_id} = req.params;
     const {list_name} = req.body;
-    const {unique_id}= req.user;
+    const {id}= req.user;
     if(!list_name ) {
         return res.status(400).json({msg: "fill the required fields"});
     }
@@ -14,18 +15,23 @@ const AddList= async(req,res,next)=>{
     }
     try{
         const board = await boardSchema.findById(board_id);
-        if(unique_id.toString() !== board.owner.toString()){
-            return res.status(400).send("Unauthorized");
-        }
         if(!board){
             return res.status(400).send("Board not found");
         }
-        const list = new listSchema({name:list_name,owner:unique_id,board:board_id,order:board.lists.length});
-        await list.save();
-        board.lists.push(list._id);
-        await board.save();
-        res.status(200).json({msg:"List added"});
+        if(unique_id.toString() !== board.owner.toString()){
+            return res.status(400).send("Unauthorized");
+        }
 
+        const lastList = await listSchema.findOne({board:board_id}).sort({order:-1}).select("order");
+        let new_order ;
+        if(!lastList) new_order='a';
+        else {
+            new_order=order_resolver(lastList.order,"");
+        }
+        if(!new_order) return res.status(500).json({msg:"internal server error"});
+        const list = new listSchema({name:list_name,owner:id,board:board_id,order:new_order});
+        await list.save();
+        res.status(200).json({msg:"List added"});
     }        
     catch(err){
         console.log("Error occurred while adding list:", err);
@@ -35,16 +41,19 @@ const AddList= async(req,res,next)=>{
 const UpdateList = async(req,res,next)=>{
     const {list_id} = req.params;
     const {list_name} = req.body;
+    const {id} = req.user;
     if(!list_id || !list_name){
         return res.status(400).json({
             msg: "fill all the required fields"
         })
     }
     try{
+        
         const list = await listSchema.findById(list_id);
         if(!list){
             return res.status(400).send("List not found");
         }
+        if(list.owner.toString()!== id) return res.status(400).json({msg: "unauthorized"});
         list.name = list_name;
         await list.save();
         res.status(200).json({msg:"List updated"});
@@ -56,16 +65,16 @@ const UpdateList = async(req,res,next)=>{
 }
 const DeleteList = async(req,res,next)=>{
     const {list_id} = req.params;
+    const {id} = req.user;
     if(!list_id) return res.status(400).json({msg:"select a list to delete"})
     try{
         const list = await listSchema.findById(list_id);
         if(!list){
             return res.status(400).send("List not found");
         }
-        const tasks = await taskSchema.find({list:list_id});
-        await Promise.all(tasks.map(async (task) => {
-            await commentSchema.deleteMany({task:task._id});
-        }))
+        if(list.owner.toString()!== id) return res.status(400).json({msg:"forbidden"});
+        const tasks = await taskSchema.find({list:list_id}).select("_id");
+        await commentSchema.deleteMany({task:{$in: tasks}});
         await taskSchema.deleteMany({list:list_id});
         await listSchema.deleteOne({_id:list_id});
         res.status(200).json({msg:"List deleted"});
@@ -77,9 +86,9 @@ const DeleteList = async(req,res,next)=>{
 }
 const ChangeOrder = async(req,res,next)=>{
     const {list_id} = req.params;
-    const {new_order} = req.body;
+    const {LB,TB} = req.body;
     
-    if(!list_id || !new_order) return res.status(400).json({
+    if(!list_id || (!LB && !TB)) return res.status(400).json({
         msg: "fill the requried fields"
     })
     try{
@@ -88,15 +97,11 @@ const ChangeOrder = async(req,res,next)=>{
         if(!list){
             return res.status(400).send("List not found");
         }
-        const list2 = await listSchema.findOne({order:new_order});
-        if(!list2){
-            return res.status(400).send("Order should be less than total number of lists");
-        }
-        const temp = list.order;
-        list.order = list2.order;
-        list2.order = temp;
+       
+        const new_order = order_resolver(LB,TB);
+        if(new_order) return res.status(500).send("internal server error");
+        list.order = new_order;
         await list.save();
-        await list2.save();
         res.status(200).json({msg:"List order updated"});
     }
     catch(err){
@@ -104,22 +109,23 @@ const ChangeOrder = async(req,res,next)=>{
         res.status(500).send("Internal Server Error");
     }
 }
-const getAllTasks = async(req,res,next)=>{
-    const {list_id} = req.params;
+
+const getLists = async(req,res,next)=>{
+    const {board_id} = req.params;
     const {page} = req.query;
-    if(!list_id) return res.status(400).json({msg:"select a list to get tasks"});
-    try{
-        const list = await listSchema.findById(list_id);
-        if(!list){
-            return res.status(400).send("List not found");
+    if(!board_id) return res.status(400).json({msg:"select a board to get lists"})
+    try {
+        const board = await boardSchema.findById(board_id);
+        if(!board){
+            return res.status(400).send("Board not found");
         }
-        const tasks = await taskSchema.find({list:list_id}).skip((page)*10).limit(10).sort({order:1});
-        res.status(200).json(tasks);
-    }
-    catch(err){
-        console.log("Error occurred while getting tasks:", err);
+        const lists= await  listSchema.find({board:board_id}).skip((page)*10).limit(10).sort({order:1});
+        
+    res.status(200).json({lists: lists});
+    } catch (error) {
+        console.log("Error occurred while getting lists:", error);
         res.status(500).send("Internal Server Error");
     }
 }
 
-export default {AddList,UpdateList,DeleteList,ChangeOrder,getAllTasks};
+export default {AddList,UpdateList,DeleteList,ChangeOrder,getLists};
