+++ backend/src/controllers/task.controller.js
@@ -1,9 +1,10 @@
 import taskSchema from "../models/taskSchema.js"
 import listSchema from "../models/listSchema.js"
 import userSchema from "../models/userSchema.js";
-
+import commentSchema from "../models/commentSchema.js"
 const createTask = async(req,res,next)=>{
       const {list_id} = req.params;
+      const{id}=req.user;
       const {task_name,task_description,task_deadline} = req.body;
       if(!list_id || !task_name || !task_description || !task_deadline){
           return res.status(400).json({
@@ -15,11 +16,11 @@ const createTask = async(req,res,next)=>{
           if(!list){
               return res.status(400).send("List not found");
           }
+          if(list.owner.toString()!==id) return res.status(400).send("Unauthorized");
+
           const deadline = new Date(task_deadline);
-          const task = new taskSchema({name:task_name,list:list_id,owner:list.owner,order:list.tasks.length,description:task_description,deadline:deadline,board:list.board});
+          const task = new taskSchema({name:task_name,list:list_id,owner:list.owner,description:task_description,deadline:deadline});
           await task.save();
-          list.tasks.push(task._id);
-          await list.save();
           res.status(200).json({msg:"Task added"});
       }
       catch(err){
@@ -35,7 +36,7 @@ const taskDetails = async(req,res,next)=>{
           })
       }
       try{
-          const task = await taskSchema.findById(task_id);
+          const task = await taskSchema.findById(task_id).populate("collabrators","name email").populate("owner","name email");
           if(!task){
               return res.status(400).send("Task not found");
           }
@@ -51,7 +52,6 @@ const updateTask = async(req,res,next)=>{
       const task_name = req.body.task_name;
       const task_description=req.body.task_description;
       const task_deadline=req.body.task_deadline;
-      const task_order=req.body.task_order;
       if(!task_id) return res.status(400).json({msg:"select a task to update"})
 
       try{
@@ -65,7 +65,6 @@ const updateTask = async(req,res,next)=>{
           if(task_name) task.name = task_name;
          if(task_description) task.description = task_description;
          if(task_deadline) task.deadline = task_deadline;
-         if(task_order) task.order = task_order;
           await task.save();
           res.status(200).json({msg:"Task updated"});
       }
@@ -91,16 +90,16 @@ const taskStatusUpdate=async(req,res,next)=>{
 }
 const deleteTask = async(req,res,next)=>{
       const{task_id} = req.params;
+      const {id}=req.user;
       if(!task_id) return res.status(400).json({msg:"select a task to delete"});
       try{
           const task = await taskSchema.findById(task_id);
           if(!task){
               return res.status(400).send("Task not found");
           }
-          const list = await listSchema.findById(task.list);
-          list.tasks.pull(task_id);
+           if(task.owner.toString()!==id) return res.status(400).send("unauthorized");
+          await commentSchema.deleteMany({task:task._id})
           await taskSchema.deleteOne({_id:task_id});
-          await list.save();
           res.status(200).json({msg:"Task deleted"});
       }
       catch(err){
@@ -110,27 +109,25 @@ const deleteTask = async(req,res,next)=>{
 }
 const assignTask = async(req,res,next)=>{
       const {task_id} = req.params;
-      const {unique_id} = req.user;
-      const {user_id} = req.body;
+      const {id} = req.user;
+      const {email} = req.body;
       if(!task_id) return res.staus(400).json({msg:"select a task to assign"});
       try{
           const task = await taskSchema.findById(task_id);
           if(!task){
               return res.status(400).send("Task not found");
           }
-          if(task.owner.toString() !== unique_id){
+          if(task.owner.toString() !== id){
               return res.status(400).send("Unauthorized");
           }
-          const collabrator=await userSchema.findOne({unique_id:user_id});
+          const collabrator=await userSchema.findOne({email});
           if(!collabrator){
               return res.status(400).send("Collabrator not found");
           }
-          if(task.collabrators.includes(user_id)){
+          if(task.collabrators.includes(collabrator._id)){
               return res.status(400).json({msg:"Task already assigned"});
           }
-          task.collabrators.push(user_id);
-          collabrator.boards.push(task.board);
-          await collabrator.save();
+          task.collabrators.push(collabrator._id);
           await task.save();
           res.status(200).json({msg:"Task assigned"});
       }
@@ -139,28 +136,22 @@ const assignTask = async(req,res,next)=>{
           res.status(500).send("Internal Server Error");
       }
 }
-const changeTaskOrder = async(req,res,next)=>{
-      const {task_id} = req.params;
-      const {order} = req.body;
-      if(!task_id) return res.status(400).json({msg:"select a task to change order"});
-      try{
-          const task = await taskSchema.findById(task_id);
-          if(!task){
-              return res.status(400).send("Task not found");
-          }
-          const task2= await taskSchema.findOne({order:order});
-          if(!task2) return res.status(400).json({msg:"Order should be less than total number of tasks"});
-          const temp = task.order;
-          task.order = task2.order;
-          task2.order = temp;
-          await task.save();
-          await task2.save();
-          res.status(200).json({msg:"Task order changed"});
-      }
-      catch(err){
-          console.log("Error occurred while changing task order:", err);
-          res.status(500).send("Internal Server Error");
-      }
+const getAllTasks = async(req,res,next)=>{
+    const {list_id} = req.params;
+    const {page} = req.query;
+    if(!list_id) return res.status(400).json({msg:"select a list to get tasks"});
+    try{
+        const list = await listSchema.findById(list_id);
+        if(!list){
+            return res.status(400).send("List not found");
+        }
+        const tasks = await taskSchema.find({list:list_id}).skip((page)*10).limit(10).sort({order:1});
+        res.status(200).json(tasks);
+    }
+    catch(err){
+        console.log("Error occurred while getting tasks:", err);
+        res.status(500).send("Internal Server Error");
+    }
 }
 
-export default {createTask,taskDetails,updateTask,taskStatusUpdate,deleteTask,assignTask,changeTaskOrder};
+export default {createTask,taskDetails,updateTask,taskStatusUpdate,deleteTask,assignTask,getAllTasks};
